data-kernel: EMAC: Disable EEE in link down to fix link toggle issue
PHY link status in EEE is not updated for link down and PHY init failure.
This results in link toggle issue for 100Mbps mode.
Fix EEE state in link down and  phy init failure to fix the toggle issue.
Disable EEE in if local PHY does not have support.

Change-Id: If164b9197a942728f7af3980e76a2b051be6f25a
CRs-Fixed: 2231214
Acked-by: Rahul Kawadgave <rahulak@qti.qualcomm.com>
Signed-off-by: Sunil Paidimarri <hisunil@codeaurora.org>