joyeuse_defconfig: Allow compiler to uninline functions marked 'inline'
This option determines if the kernel forces gcc to inline the functions
developers have marked 'inline'. Doing so takes away freedom from gcc to
do what it thinks is best, which is desirable for the gcc 3.x series of
compilers. The gcc 4.x series have a rewritten inlining algorithm and
enabling this option will generate a smaller kernel there. Hopefully
this algorithm is so good that allowing gcc 4.x and above to make the
decision will become the default in the future. Until then this option
is there to test gcc for this.

Change-Id: If2f7c5c35a949e9c38d4ab68cc41ecbccabb30dd
Signed-off-by: Panchajanya1999 <panchajanya@azure-dev.live>